+++ src/ElementTemplateChooser.js
@@ -10,18 +10,21 @@ import './ElementTemplateChooser.css';
  * @param {ElementTemplates} elementTemplates
  * @param {Translate} translate
  * @param {PopupMenu} popupMenu
+ * @param {Canvas} canvas
  */
 export default function ElementTemplateChooser(
     config,
     eventBus,
     elementTemplates,
     translate,
-    popupMenu) {
+    popupMenu,
+    canvas) {
 
   this._eventBus = eventBus;
   this._elementTemplates = elementTemplates;
   this._translate = translate;
   this._popupMenu = popupMenu;
+  this._canvas = canvas;
 
   const enableChooser = !config || config.elementTemplateChooser !== false;
 
@@ -44,7 +47,8 @@ ElementTemplateChooser.$inject = [
   'eventBus',
   'elementTemplates',
   'translate',
-  'popupMenu'
+  'popupMenu',
+  'canvas'
 ];
 
 ElementTemplateChooser.prototype.open = function(element) {
@@ -52,10 +56,22 @@ ElementTemplateChooser.prototype.open = function(element) {
   const popupMenu = this._popupMenu;
   const translate = this._translate;
   const eventBus = this._eventBus;
+  const canvas = this._canvas;
+
+  const restoreCanvasFocus = () => {
+
+    // Only available with diagram-js >= 15.0.0
+    if (canvas && canvas.restoreFocus) {
+      canvas.restoreFocus();
+    }
+  };
 
   return new Promise((resolve, reject) => {
 
-    const handleClosed = () => reject('user-canceled');
+    const handleClosed = () => {
+      reject('user-canceled');
+      restoreCanvasFocus();
+    };
 
     eventBus.once('popupMenu.close', handleClosed);
 
@@ -66,6 +82,7 @@ ElementTemplateChooser.prototype.open = function(element) {
       eventBus.off('popupMenu.close', handleClosed);
 
       resolve(template);
+      restoreCanvasFocus();
     });
 
     popupMenu.open(element, 'element-template-chooser', { x: 0, y: 0 }, {